requirements_additions.txt
# Additional dependencies for product routes
requests>=2.28.0
flask-cors>=4.0.0
sqlalchemy>=1.4.0
